Added method to update the Lambda functions fixed slack_message lambda to use the new pip apis

+++ pbx_gs_python_utils/utils/Files.py
@@ -77,8 +77,10 @@ class Files:
 
     @staticmethod
     def folder_delete_all(path):                # this will remove recursively
-        shutil.rmtree(path)
-        return Files.exists(path) is False
+        if Files.folder_exists(path):
+            shutil.rmtree(path)
+            return Files.exists(path) is False
+        return True
 
     @staticmethod
     def folder_name(path):
